"Implemented database connectivity with Hibernate"
In the applicationContext.xml, a data source has been set up to connect to a PostgreSQL database. Combined with the added transactions should significantly improve application performance by managing pooled JDBC connections. The Pom.xml has been updated to include the latest version of PostgreSQL and Hibernate as well as a number of additional dependencies to facilitate connection pooling and transaction management. It has also been set up to scan the appropriate packages.

Meanwhile, the web.xml was adjusted to reflect the changes in the application context and to update the display name of the application.