"Improve debugging and display of Employee data"
Introduced console logging for Employee data in MyController class for improved debugging. The data fetched from the database are now printed in the console. The all-employees-view.jsp was also refactored. The Employee attributes are now displayed using standard getter methods, replacing the less efficient use of <jsp:useBean> which was causing bugs.