Added "add new employee" functionality and updated views
Added a new JSP form to collect employee information and created the method saveEmployee() in EmployeeDAO, EmployeeService, EmployeeDAOImpl and EmployeeServiceImpl. Extended the MyController class to handle the new functionalities.

Also, added an "Add employee" button onto all-employees-view.jsp to navigate user to the employee input form. These changes provide a way for users to add a new employee to the system improving the overall functionality and user-experience of the web application.