Remove commented EmployeeDAO dependency and clarify form actions
Removed an unnecessary commented dependency (EmployeeDAO) from the MyController class, making the system rely on the EmployeeService only, hence simplifying the class.

Additionally, added comments in both the employee-info.jsp and all-employees-view.jsp to clarify the action of the forms. In employee-info.jsp, a comment was added to describe the action upon clicking submit. Similarly, in the all-employees-view.jsp, a comment was added to explain the navigation upon clicking the "add" button.