Switch to server controlled datasource and update applicationContext
This commit switches our data-source control from the application to the server. It introduces a context.xml file to declare the postgres database as a <Resource> to be managed by the server. It also modifies the applicationContext.xml file to replace the c3p0 datasource with a JndiObjectFactoryBean for using the server managed data-source. This change is aiming to optimize resource usage as the server is typically more efficient in managing database connections than a standalone application. A relevant comment in the applicationContext.xml file has also been updated to reflect this change.