Add unit tests for MyController and update test dependencies
This commit adds unit tests for the MyController class to ensure the functionality of its methods. The tests are implemented using JUnit and Mockito, which are also added as dependencies in the pom file. An unnecessary @Transactional annotation in the EmployeeService class is also removed. Additionally, a 'proxyInterface' property is added to the 'dataSource' bean in the applicationContext file, allowing us to make use of Tomcat's connection pooling mechanism. This makes the application more efficient by reusing database connections instead of creating a new one for every request.